+++ tests/KernelTest.php
@@ -2,13 +2,13 @@
 
 namespace Tapestry\Tests;
 
-use Symfony\Component\Finder\SplFileInfo;
-use Tapestry\Entities\ContentType;
-use Tapestry\Entities\File;
-use Tapestry\Modules\ContentTypes\ContentTypeFactory;
+use Tapestry\Providers\ProjectKernelServiceProvider;
+use Tapestry\Tests\Traits\MockTapestry;
 
 class KernelTest extends CommandTestBase
 {
+
+    use MockTapestry;
 
     /**
      * Written for issue #133
@@ -64,5 +64,21 @@ class KernelTest extends CommandTestBase
 
         $this->assertEquals('Hello world! This command was loaded via a site Kernel.', trim($output->getDisplay()));
         $this->assertEquals(0, $output->getStatusCode());
+    }
+
+    /**
+     * Written for issue #234
+     * @link https://github.com/carbontwelve/tapestry/issues/234
+     * @expectedException \Exception
+     */
+    public function testKernelThrowsException()
+    {
+        $this->copyDirectory('assets/build_test_37/src', '_tmp');
+        $class = new ProjectKernelServiceProvider();
+        $tapestry = $this->mockTapestry(__DIR__ . DIRECTORY_SEPARATOR . '_tmp');
+        $class->setContainer($tapestry->getContainer());
+
+        $this->expectExceptionMessage('[SiteThirtySeven\Kernel] kernel file not found.');
+        $class->boot();
     }
 }